pyproject.toml
[tool.poetry]
name = "seek-for-humanoid-server"
version = "0.1.0"
description = ""
authors = ["Nicolas Calvello <nic.calv@outlook.it>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
djangorestframework = "^3.12.4"
Pillow = "^8.4.0"
requests = "^2.26.0"
django-cors-headers = "^3.10.0"
gunicorn = "^20.1.0"
django-storages = "^1.12.3"
boto3 = "^1.19.7"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
